requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
chromadb==0.4.17
colorama==0.4.6
openai==1.59.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pypdf>=5.1.0
pdfplumber>=0.11.5
uvicorn>=0.34.0
pydantic>=2.9.2
fastapi>=0.115.6